+++ src/Plugin.php
@@ -136,7 +136,7 @@ class Plugin extends BasePlugin
     /**
      * @inheritdoc
      */
-    public $minVersionRequired = '1.2.1360';
+    public $minVersionRequired = '2.2.17';
 
     use CommerceServices;
     use Variables;