+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Release Notes for Craft Commerce
 
+## Unreleased
+
+### Added
+- Added `validateCustomFields` form parameter to allow custom field validation when posting to `commerce/cart/update-cart`. ([#1292](https://github.com/craftcms/commerce/issues/1292))
+
+### Changed
+- Updated the minimum required version to upgrade to `2.2.17`.
+
 ## 3.0.11 - 2020-02-25
 
 ### Added